+++ lib/test_prof/rspec_dissect.rb
@@ -7,6 +7,9 @@ module TestProf
   # RSpecDissect tracks how much time do you spend in `before` hooks
   # and memoization helpers (i.e. `let`) in your tests.
   module RSpecDissect
+    class Span < Struct.new(:id, :parent_id, :type, :duration, :meta)
+    end
+
     module ExampleInstrumentation # :nodoc:
       def run_before_example(*)
         RSpecDissect.track(:before) { super }
@@ -15,12 +18,15 @@ module TestProf
 
     module MemoizedInstrumentation # :nodoc:
       def fetch_or_store(id, *)
+        return super if id == :subject
+        return @memoized[id] if @memoized[id]
+
         res = nil
         Thread.current[:_rspec_dissect_let_depth] ||= 0
         Thread.current[:_rspec_dissect_let_depth] += 1
         begin
           res = if Thread.current[:_rspec_dissect_let_depth] == 1
-            RSpecDissect.track(:let, id) { super }
+            RSpecDissect.track(:let, name: id) { super }
           else
             super
           end
@@ -33,44 +39,24 @@ module TestProf
 
     # RSpecDisect configuration
     class Configuration
-      MODES = %w[all let before].freeze
-
       attr_accessor :top_count, :let_stats_enabled,
         :let_top_count
 
       alias_method :let_stats_enabled?, :let_stats_enabled
-
-      attr_reader :mode
 
       def initialize
         @let_stats_enabled = true
         @let_top_count = (ENV["RD_PROF_LET_TOP"] || 3).to_i
         @top_count = (ENV["RD_PROF_TOP"] || 5).to_i
         @stamp = ENV["RD_PROF_STAMP"]
-        @mode = (ENV["RD_PROF"] == "1") ? "all" : ENV["RD_PROF"]
-
-        unless MODES.include?(mode)
-          raise "Unknown RSpecDissect mode: #{mode};" \
-                "available modes: #{MODES.join(", ")}"
-        end
 
         RSpecStamp.config.tags = @stamp if stamp?
-      end
-
-      def let?
-        mode == "all" || mode == "let"
-      end
-
-      def before?
-        mode == "all" || mode == "before"
       end
 
       def stamp?
         !@stamp.nil?
       end
     end
-
-    METRICS = %w[before let].freeze
 
     class << self
       include Logging
@@ -89,56 +75,64 @@ module TestProf
         RSpec::Core::MemoizedHelpers::ThreadsafeMemoized.prepend(MemoizedInstrumentation)
         RSpec::Core::MemoizedHelpers::NonThreadSafeMemoized.prepend(MemoizedInstrumentation)
 
-        @data = {}
-
-        METRICS.each do |type|
-          @data["total_#{type}"] = 0.0
-        end
-
         reset!
 
         log :info, "RSpecDissect enabled"
       end
 
-      def track(type, meta = nil)
-        start = TestProf.now
-        res = yield
-        delta = (TestProf.now - start)
-        type = type.to_s
-        @data[type][:time] += delta
-        @data[type][:meta] << meta unless meta.nil?
-        @data["total_#{type}"] += delta
-        res
+      def nextid
+        @last_id += 1
+        @last_id.to_s
       end
 
-      def reset!
-        METRICS.each do |type|
-          @data[type.to_s] = {time: 0.0, meta: []}
+      def current_span
+        Thread.current[:_rspec_dissect_spans_stack].last
+      end
+
+      def span_stack
+        Thread.current[:_rspec_dissect_spans_stack]
+      end
+
+      def track(type, id: nextid, **meta)
+        span = Span.new(id, current_span&.id, type, 0.0, meta)
+        span_stack << span
+
+        begin
+          start = TestProf.now
+          res = yield
+          delta = (TestProf.now - start)
+          span.duration = delta
+          @spans << span
+          res
+        ensure
+          span_stack.pop
         end
       end
 
-      # Whether we are able to track `let` usage
-      def memoization_available?
-        defined?(::RSpec::Core::MemoizedHelpers::ThreadsafeMemoized)
+      def populate_from_spans!(data)
+        data[:total_setup] = @spans.select { !_1.parent_id }.sum(&:duration)
+        data[:total_before_let] = @spans.select { _1.type == :let && _1.parent_id }.sum(&:duration).to_f
+        data[:total_lazy_let] = @spans.select { _1.type == :let && !_1.parent_id }.sum(&:duration).to_f
+
+        data[:top_lets] = @spans.select { _1.type == :let }
+          .group_by { _1.meta[:name] }
+          .transform_values! do |spans|
+            {name: spans.first.meta[:name], duration: spans.sum(&:duration), size: spans.size}
+          end
+          .values
+          .sort_by { -_1[:duration] }
+          .take(RSpecDissect.config.let_top_count)
       end
 
-      def time_for(key)
-        @data[key.to_s][:time]
-      end
-
-      def meta_for(key)
-        @data[key.to_s][:meta]
-      end
-
-      def total_time_for(key)
-        @data["total_#{key}"]
+      def reset!
+        @last_id = 1
+        @spans = []
+        Thread.current[:_rspec_dissect_spans_stack] = []
       end
     end
   end
 end
 
-require "test_prof/rspec_dissect/collectors/let"
-require "test_prof/rspec_dissect/collectors/before"
 require "test_prof/rspec_dissect/rspec"
 
 TestProf.activate("RD_PROF") do