+++ spec/integrations/rspec_dissect_spec.rb
@@ -8,77 +8,13 @@ describe "RSpecDissect" do
 
     expect(output).to include("RSpecDissect report")
     expect(output).to match(/Total time:\s+\d{2}:\d{2}\.\d{3}/)
-    expect(output).to match(/Total `before\(:each\)` time:\s+\d{2}:\d{2}\.\d{3}/)
+    expect(output).to match(/Total setup time:\s+\d{2}:\d{2}\.\d{3}/)
 
     expect(output).to include_lines(
-      "Top 5 slowest suites (by `before(:each)` time):",
-      "Subject + Before (./rspec_dissect_fixture.rb:22) – ",
-      "Only let (./rspec_dissect_fixture.rb:43) – "
-    )
-
-    if TestProf::RSpecDissect.memoization_available?
-      expect(output).to match(/Total `let` time:\s+\d{2}:\d{2}\.\d{3}/)
-      expect(output).to include_lines(
-        "Top 5 slowest suites (by `let` time):",
-        "Only let (./rspec_dissect_fixture.rb:43) – ",
-        " ↳ work – 1",
-        " ↳ more_work – 1",
-        "Subject + Before (./rspec_dissect_fixture.rb:22) – ",
-        " ↳ work – 2",
-        " ↳ no_work – 1"
-      )
-    else
-      expect(output).to include("`let` profiling is not supported (requires RSpec >= 3.3.0)")
-    end
-  end
-
-  specify "it works with specified top count", :aggregate_failures do
-    output = run_rspec("rspec_dissect", env: {"RD_PROF" => "1", "RD_PROF_TOP" => "1"})
-
-    expect(output).to include("0 failures")
-
-    expect(output).to include_lines(
-      "Top 1 slowest suites (by `before(:each)` time):",
+      "Top 5 slowest suites by setup time:",
+      "Only let (./rspec_dissect_fixture.rb:43) – ",
       "Subject + Before (./rspec_dissect_fixture.rb:22) – "
     )
-
-    if TestProf::RSpecDissect.memoization_available?
-      expect(output).to include_lines(
-        "Top 1 slowest suites (by `let` time):",
-        "Only let (./rspec_dissect_fixture.rb:43) – "
-      )
-    end
-  end
-
-  if TestProf::RSpecDissect.memoization_available?
-    specify "it works when mode is before", :aggregate_failures do
-      output = run_rspec("rspec_dissect", env: {"RD_PROF" => "before", "RD_PROF_TOP" => "1"})
-
-      expect(output).to include("0 failures")
-
-      expect(output).to include_lines(
-        "Top 1 slowest suites (by `before(:each)` time):",
-        "Subject + Before (./rspec_dissect_fixture.rb:22) – "
-      )
-
-      expect(output).not_to include(
-        "Top 1 slowest suites (by `let` time):"
-      )
-    end
-
-    specify "it works when mode is let", :aggregate_failures do
-      output = run_rspec("rspec_dissect", env: {"RD_PROF" => "let", "RD_PROF_TOP" => "1"})
-
-      expect(output).to include("0 failures")
-
-      expect(output).not_to include(
-        "Top 1 slowest suites (by `before(:each)` time):"
-      )
-
-      expect(output).to include_lines(
-        "Top 1 slowest suites (by `let` time):"
-      )
-    end
   end
 
   context "with RStamp" do
@@ -98,13 +34,13 @@ describe "RSpecDissect" do
     specify "it works", :aggregate_failures do
       output = run_rspec(
         "rspec_dissect_stamp",
-        env: {"RD_PROF" => "1", "RD_PROF_STAMP" => "slow", "RD_PROF_TOP" => "1"}
+        env: {"RD_PROF" => "1", "RD_PROF_STAMP" => "slow_setup", "RD_PROF_TOP" => "1"}
       )
 
       expect(output).to include("5 examples, 0 failures")
 
       expect(output).to include_lines(
-        "Top 1 slowest suites (by `before(:each)` time):",
+        "Top 1 slowest suites by setup time:",
         "Subject + Before (./rspec_dissect_stamp_fixture.rb:22) – "
       )
 
@@ -116,7 +52,7 @@ describe "RSpecDissect" do
 
       output2 = run_rspec(
         "rspec_dissect_stamp",
-        env: {"SPEC_OPTS" => "--tag slow"}
+        env: {"SPEC_OPTS" => "--tag slow_setup"}
       )
 
       expect(output2).to include("3 examples, 0 failures")